Decompose RunTestDefinitions
